backend/app/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
python-multipart==0.0.9

# Google Gemini AI
google-generativeai==0.7.2
langchain==0.2.6
langchain-google-genai==1.0.6
langchain-community==0.2.6

# Voice — Deepgram speech-to-text
deepgram-sdk==3.4.0

# Vector database (knowledge base)
chromadb==0.5.3

# Database
sqlalchemy==2.0.31
aiosqlite==0.20.0

# Config & environment
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Utilities
httpx==0.27.0
aiofiles==23.2.1